requirements.txt
numpy
tensorflow
progressbar2
h5py
